+++ packages/plugin-openapi/test/fixtures/api-users-collection-from/openapi/api.js
@@ -1,0 +1,12 @@
+const openApiDocument = require("../../../openapi/users");
+
+module.exports = [
+  {
+    basePath: "/api",
+    collection: {
+      id: "users",
+      from: "base",
+    },
+    document: openApiDocument,
+  },
+];